chore: created shared type file
