+++ src/types/index.ts
@@ -1,0 +1,8 @@
+export enum ChangeType
+{ // eslint-disable-line @typescript-eslint/indent
+  NONE = "none",
+  INSERT = "insert",
+  UPDATE = "update",
+  DELETE = "delete",
+  PENDING = "pending"
+}
